initial commit
Features of iLogger
Include Metadata: The library allows you to include additional metadata in your logs, such as timestamps, thread IDs, task IDs, and instance names. This is valuable for debugging and auditing.

Variable Change Tracking: With iLogger, you can directly track changes to specific variables in your code, and these changes will automatically be logged.

Namespace-Based Logging: One of the unique features is the ability to target logging based on namespaces. This is especially useful for larger projects where you might want to focus your debugging efforts on a specific portion of your codebase.

Instance-Level Logging: The library allows you to either focus on specific instances of objects for logging or generalize to all instances of an object. This offers a lot of flexibility in terms of what gets logged.

Custom Logging Criteria: You can define custom acceptance criteria that decide when an object's state changes should be logged. These criteria can be based on the object’s current state or other dynamic factors.

Multiple Acceptance Criteria: You can set multiple conditions that an object must meet to be logged. This allows you to fine-tune what you want to appear in the logs with even more granularity.